refactor(html): tree construction "select"
