+++ html/parser/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "resworb-html-parser"
-version = "0.1.135"
+version = "1.1.136"
 license-file = "../../LICENSE"
 edition = "2021"
